+++ src/Client.ts
@@ -5,7 +5,7 @@ import { ObjectValues } from "./Types/HelperTypes.js";
 import { GatewayIntents, GatewayOpCodes } from "./Types/DiscordGateway.js";
 import { GuildCache } from "./Managers/Guilds.js";
 import { EventEmitter } from "node:events";
-import type { ClientEventMap } from "./Types/SimplicityTypes.js";
+import type { ClientEventMap } from "./Types/SimplyJSTypes.js";
 import { User } from "./Structures/User.js";
 import { UserCache } from "./Managers/Users.js";
 import { ActivityType, ClientActivity, Status } from "./Types/DiscordAPITypes.js";