+++ src/Events/Messages.ts
@@ -2,7 +2,7 @@ import { defineEvent, JSONObject } from "../Types/Internal.js";
 import { GatewayEvents } from "../Types/DiscordGateway.js";
 import { DiscordMessage } from "../Types/MessageComponents.js";
 import { Message } from "../Structures/Message.js";
-import { ClientEvents } from "../Types/SimplicityTypes.js";
+import { ClientEvents } from "../Types/SimplyJSTypes.js";
 
 export const MessageCreate = defineEvent({
 	name: GatewayEvents.MessageCreate,